Organize oreboot project into cargo workspace
Previously, oreboot consists of individual cargo packages. This would comply
with C code, but it can't share incremental build caches. In this commit we
will organize it into one workspace for cargo, in this case cargo would use
the same target path so compilation caches would be shared.

To orgainze workspace, we rearranged soc and cpu modules, as is done in driver
module. This rearrangement does not include any change to working code except
only for changes to dependency path.

This commit includes changes to build.rs file, it's mostly used in FSP build
and link script discovery. Changes to them will allow to detect the 3rdparty
directory and search for link.ld in every path of organized workspace, so
we may build them wherever we are at in the project directory.

Some Cargo.toml file is updated to unify the dependency crate version. This
change would avoid multiple version crates coexist in compile output, resulting
in oversized blobs which would overflow flash memory size. File fixed-dtfs.dts
and mainboard makefile are also modified to avoid overlap of same file names
when compiling, as well as size report files.

The commit also includes multiple clippy fixes. The project after this commit
would still need makefiles to help building, but this organization to workspace
would furtherly allow us to develop a xtask based building framework onto
oreboot project.

+++ src/drivers/src/uart/sifive.rs
@@ -101,7 +101,7 @@ impl SiFive {
 impl Driver for SiFive {
     fn init(&mut self) -> Result<()> {
         // Disable UART interrupts.
-        self.ie.set(0 as u32);
+        self.ie.set(0_u32);
         // Set clock rate to the default 33.33MHz.
         self.set_clock_rate(33330000);
         // Enable transmit.
